+++ R/rdq.test.R
@@ -3,7 +3,7 @@
 #' \code{rdq.test} provides testing results for hypotheses on the treatment effects concerning (i) treatment significance, (ii) homogeneity of effects over quantiles,
 #' and (iii) positive or negative dominance hypothesis.
 #'
-#' @usage rdq.test(y,x,d,x0,z0=NULL,tau,bdw,cov,bias,alpha,type,std.opt=1,print.qte=1)
+#' @usage rdq.test(y,x,d,x0,z0=NULL,tau,bdw,cov,bias,alpha,type,std.opt=1)
 #'
 #' @param y a numeric vector, the outcome variable.
 #' @param x a vector (or a matrix) of covariates, the first column is the running variable.
@@ -31,12 +31,12 @@
 #' and set \emph{type} to 4 to test the null hypothesis of uniformly non-positive treatment effects against the presence of positive effects at some quantiles.
 #' @param std.opt either 0 or 1. If \emph{std.opt=1}, the test statistic is standardized so that
 #' the variance is equalized across quantiles; if \emph{std.opt=0}, the test is not standardized.
-#' @param print.qte a logical flag specifying whether to print an outcome table.
 #' @return
 #' A list with elements:
 #' \describe{
-#' \item{te.st}{test statistics}
+#' \item{statistic}{test statistics.}
 #' \item{cr.va}{critical values.}
+#' \item{p.value}{p values.}
 #' }
 #' @export
 #'
@@ -63,7 +63,7 @@
 #' alpha=c(0.1,0.05),type=c(3,4))}
 #'
 #'
-rdq.test <- function(y,x,d,x0,z0=NULL,tau,bdw,cov,bias,alpha,type,std.opt=1,print.qte=1){
+rdq.test <- function(y,x,d,x0,z0=NULL,tau,bdw,cov,bias,alpha,type,std.opt=1){
   x <- as.matrix(x)
   dz <- ncol(x)-1
   mis <- apply(apply(cbind(y,x,d),2,is.na),1,max)
@@ -82,7 +82,7 @@ rdq.test <- function(y,x,d,x0,z0=NULL,tau,bdw,cov,bias,alpha,type,std.opt=1,prin
   if(length(bdw)==1) {bdw.opt <- 1}
   if(length(bdw)>1 & length(bdw)==length(tau)) {bdw.opt <- 2}
   if(length(bdw)>1 & length(bdw)!=length(tau))
-    {stop("The length of bdw should be one or equal to the length of tau.")}
+  {stop("The length of bdw should be one or equal to the length of tau.")}
   if(bdw.opt==1){
     tt <- sort(unique(c(tau,0.5)))	# qualtile levels to estimate
     hh <- bdw*((2*tt*(1-tt)/(pi*dnorm(qnorm(tt))^{2}))^{1/5})  # quantile specific bandwidths
@@ -104,72 +104,15 @@ rdq.test <- function(y,x,d,x0,z0=NULL,tau,bdw,cov,bias,alpha,type,std.opt=1,prin
   # tests for QTE
   sm <- rdq.sim(x,d,x0,z0,dz,cov,tt=tt,hh2[ind],hh2[ind],fxp=fp$ff,fxm=fm$ff,n.sim)
   te.st <- list(); cr.va <- list()
+  p.val <- list()
   for(j in 1:length(type)){
     bt <- run.test(n,dz,taus=tt,hh2[ind],Dc.p=sm$dcp,Dc.m=sm$dcm,Dr.p=sm$drp,Dr.m=sm$drm,Qy.p=as.matrix(ab$qp.est[ind,]),Qy.m=as.matrix(ab$qm.est[ind,]),bias.p=bp$bias,bias.m=bm$bias,cov,bias,alpha,n.sim,test.type=type[j],std.opt)
-    if(type[j]==1){te.st$significance = bt$test.stat; cr.va$significance = bt$cr.value}
-    if(type[j]==2){te.st$homogeneity = bt$test.stat; cr.va$homogeneity = bt$cr.value}
-    if(type[j]==3){te.st$unambiguity = bt$test.stat; cr.va$unambiguity = bt$cr.value}
-    if(type[j]==4){te.st$ne.unambiguity = bt$test.stat; cr.va$ne.unambiguity = bt$cr.value}
+    if(type[j]==1){te.st$significance = bt$test.stat; cr.va$significance = bt$cr.value; p.val$significance = bt$p.val}
+    if(type[j]==2){te.st$homogeneity = bt$test.stat; cr.va$homogeneity = bt$cr.value; p.val$homogeneity = bt$p.val}
+    if(type[j]==3){te.st$unambiguity = bt$test.stat; cr.va$unambiguity = bt$cr.value; p.val$unambiguity = bt$p.val}
+    if(type[j]==4){te.st$ne.unambiguity = bt$test.stat; cr.va$ne.unambiguity = bt$cr.value; p.val$ne.unambiguity = bt$p.val}
   }
-  if(print.qte==1){
-    cat("\n\n")
-    cat(format("Testing hypotheses on quantile process",
-               width = 75, justify = "centre"), "\n")
-    cat(paste(rep("-", 75), collapse = ""), "\n")
-    cat(format("NULL Hypthoesis", width = 38), "\t",
-        format("test stat.", width = 12, justify = "centre"),
-        format("critical value", width = 12, justify = "centre"), "\n")
-    cat(format("",width = 55, collapse = ""),
-        format(paste((100*alpha),"%","      ",sep=""), justify = "left"),"\n")
-    cat(paste(rep("=", 75), collapse = ""), "\n")
-    if(cov==0){
-      if(1 %in% type){
-        cat(format("Significance: QTE(tau|x,z)=0 for all taus ", width = 42),
-            format(te.st$significance, digits = 2, nsmall = 2, width = 8),
-            format(cr.va$significance, digits = 2, nsmall = 2, width = 8), "\n")
-      }
-      if(2 %in% type){
-        cat(format("Homogeneity: QTE(tau|x,z) is constant ", width = 42),
-            format(te.st$homogeneity, digits = 2, nsmall = 2, width = 8),
-            format(cr.va$homogeneity, digits = 2, nsmall = 2, width = 8), "\n")
-      }
-      if(3 %in% type){
-        cat(format("Dominance: QTE(tau|x,z)>=0 for all taus ", width = 42),
-            format(te.st$unambiguity, digits = 2, nsmall = 2, width = 8),
-            format(cr.va$unambiguity, digits = 2, nsmall = 2, width = 8), "\n")
-      }
-      if(4 %in% type){
-        cat(format("Dominance: QTE(tau|x,z)<=0 for all taus ", width = 42),
-            format(te.st$ne.unambiguity, digits = 2, nsmall = 2, width = 8),
-            format(cr.va$ne.unambiguity,digits = 2, nsmall = 2, width = 8), "\n")
-      }
-    }
-    if(cov==1){
-      for(j in 1:dg){
-        if(j>1){cat(paste(rep("-", 75), collapse = ""), "\n")}
-        cat(format(paste("Group-",j,sep=""), width = 10,justify ="centre"),"\n", sep = "")
-        if(1 %in% type){
-          cat(format("Significance: QTE(tau|x,z)=0 for all taus ", width = 42),
-              format(te.st$significance[j,], digits = 2, nsmall = 2, width = 8),
-              format(cr.va$significance[j,], digits = 2, nsmall = 2, width = 8), "\n")
-        }
-        if(2 %in% type){
-          cat(format("Homogeneity: QTE(tau|x,z) is constant ", width = 42),
-              format(te.st$homogeneity[j,], digits = 2, nsmall = 2, width = 8),
-              format(cr.va$homogeneity[j,], digits = 2, nsmall = 2, width = 8), "\n")
-        }
-        if(3 %in% type){
-          cat(format("Dominance: QTE(tau|x,z)>=0 for all taus ", width = 42),
-              format(te.st$unambiguity[j,], digits = 2, nsmall = 2, width = 8),
-              format(cr.va$unambiguity[j,], digits = 2, nsmall = 2, width = 8), "\n")
-        }
-        if(4 %in% type){
-          cat(format("Dominance: QTE(tau|x,z)<=0 for all taus ", width = 42),
-              format(te.st$ne.unambiguity[j,], digits = 2, nsmall = 2, width = 8),
-              format(cr.va$ne.unambiguity[j,],digits = 2, nsmall = 2, width = 8), "\n")
-        }
-      }
-    }
-  }
-  return(list(test.stat = te.st, cr.value = cr.va))
+  out <- list(statistic = te.st, cr.value = cr.va, p.value = p.val, type=type, cov=cov, dg=dg, alpha=alpha)
+  class(out) <- c("test.qte", class(out))
+  return(out)
 }