+++ NEWS.md
@@ -1,3 +1,9 @@
 # QTE.RD 1.0.0
 
 * Initial CRAN submission.
+
+# QTE.RD 1.1.0
+
+* First revision.
+* added summary.qte() and several print methods.
+* added p-values in the outcome of rdq.test().